requirements.txt
alabaster==1.0.0
babel==2.17.0
certifi==2025.11.12
charset-normalizer==3.4.4
docutils==0.21.2
firebird-base==2.0.2
firebird-driver==2.0.2
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
protobuf==5.29.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
roman-numerals-py==3.1.0
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
sqlalchemy-firebird==2.1
typing_extensions==4.15.0
urllib3==2.5.0
